+++ data-raw/storms.R
@@ -4,11 +4,13 @@ library(tidyverse)
 # in an unorthodox format: a csv that alternates between header/identifier rows
 # and data rows.
 
+# TO UPDATE: get the latest URL from https://www.nhc.noaa.gov/data/#hurdat, and rerun this code
+
 # Read in data set so each line is a character string
-storm_strings <- read_lines("http://www.nhc.noaa.gov/data/hurdat/hurdat2-1851-2015-070616.txt")
+storm_file_complete <- read_file("https://www.nhc.noaa.gov/data/hurdat/hurdat2-1851-2020-052921.txt")
+storm_strings <- read_lines(storm_file_complete)
 
 # Identify the header lines that have three commas
-library(stringr)
 header_locations <- (1:length(storm_strings))[str_count(storm_strings, "\\,") == 3]
 
 # Extract length of each sub-dataset
@@ -20,47 +22,60 @@ headers_df <- headers %>%
   mutate(name = recode(name, "UNNAMED" = id), skip = header_locations) %>%
   select(name, skip, n_obs)
 
-# Read in the sub-datasets as data frames
-df_names <- c(
-  "date", "time", "record_type", "status", "lat", "long", "wind", "pressure",
-  "extent_34_NE", "extent_34_SE", "extent_34_SW", "extent_34_NW",
-  "extent_50_NE", "extent_50_SE", "extent_50_SW", "extent_50_NW",
-  "extent_64_NE", "extent_64_SE", "extent_64_SW", "extent_64_NW", "nas"
+column_types <- list(
+  date = col_character(),
+  time = col_character(),
+  record_type = col_character(),
+  status = col_character(),
+  lat = col_character(),
+  long = col_character(),
+  wind = col_integer(),
+  pressure = col_integer(),
+  extent_34_NE = col_integer(),
+  extent_34_SE = col_integer(),
+  extent_34_SW = col_integer(),
+  extent_34_NW = col_integer(),
+  extent_50_NE = col_integer(),
+  extent_50_SE = col_integer(),
+  extent_50_SW = col_integer(),
+  extent_50_NW = col_integer(),
+  extent_64_NE = col_integer(),
+  extent_64_SE = col_integer(),
+  extent_64_SW = col_integer(),
+  extent_64_NW = col_integer(),
+  nas = col_integer()
 )
+column_names <- names(column_types)
 
-storm_dfs <- vector("list", nrow(headers_df))
-names(storm_dfs) <- headers_df$name
 
-for (i in seq_along(headers_df$name)) {
-  storm_dfs[[i]] <- read_csv("data-raw/hurdat2.txt",
-    skip = headers_df$skip[i],
-    n_max = headers_df$n_obs[i],
-    col_names = df_names,
-    na = c("", "-99", "-999"),
-    col_types = list(
-      time = col_character(),
-      pressure = col_integer(),
-      extent_34_NE = col_integer(),
-      extent_34_SE = col_integer(),
-      extent_34_SW = col_integer(),
-      extent_34_NW = col_integer(),
-      extent_50_NE = col_integer(),
-      extent_50_SE = col_integer(),
-      extent_50_SW = col_integer(),
-      extent_50_NW = col_integer(),
-      extent_64_NE = col_integer(),
-      extent_64_SE = col_integer(),
-      extent_64_SW = col_integer(),
-      extent_64_NW = col_integer()
-    )
+#### Parse each storm as its own sub-dataframe
+storm_dataframes <- vector("list", nrow(headers_df))
+for (i in 1:nrow(headers_df)) {
+  # get this storm's metadata
+  row_start = headers_df[i,]$skip + 1
+  row_end = headers_df[i,]$n_obs + row_start - 1
+  # subset of rows belonging to this storm
+  data_subset = storm_strings[row_start:row_end] %>%
+    paste(collapse = "\n") %>%
+    paste0("\n")
+  data_subset = read_csv(
+    data_subset,
+    col_names = column_names,
+    col_types = column_types,
+    na = c("", "-99", "-999")
   )
+  # name at the front
+  data_subset$name = headers_df[i,]$name
+  data_subset = data_subset %>% relocate(name)
+  # add to list of storms
+  storm_dataframes[[i]] = data_subset
 }
 
 # Combine and clean the data sets
 library(lubridate)
 
-storms <- storm_dfs %>%
-  bind_rows(.id = "name") %>%
+storms <- storm_dataframes %>%
+  bind_rows() %>%
   mutate(
     date = ymd(date),
     year = year(date),
@@ -81,13 +96,13 @@ storms <- storm_dfs %>%
     # wind = wind * 1.15078, # transforms knots to mph,
     TSradius1 = extent_34_NE + extent_34_SW,
     TSradius2 = extent_34_NW + extent_34_SE,
-    ts_diameter = pmax(TSradius1, TSradius2) * 1.15078, # to convert from nautical miles to miles
+    tropicalstorm_force_diameter = pmax(TSradius1, TSradius2),
     HUradius1 = extent_64_NE + extent_64_SW,
     HUradius2 = extent_64_NW + extent_64_SE,
-    hu_diameter = pmax(HUradius1, HUradius2) * 1.15078, # to convert from nautical miles to miles
+    hurricane_force_diameter = pmax(HUradius1, HUradius2),
     status = recode(status, "HU" = "hurricane", "TS" = "tropical storm", "TD" = "tropical depression")
   ) %>%
-  select(name, year, month, day, hour, lat, long, status, category, wind, pressure, ts_diameter, hu_diameter)
+  select(name, year, month, day, hour, lat, long, status, category, wind, pressure, tropicalstorm_force_diameter, hurricane_force_diameter)
 
 # Narrow to storms that have complete pressure record
 completeish <- storms %>%
@@ -103,4 +118,5 @@ storms <- storms %>%
   ) %>%
   mutate(name = if_else(str_sub(name, 1, 3) %in% c("AL0", "AL1"), name, str_to_title(name)))
 
-devtools::use_data(storms)
+# output for the package
+usethis::use_data(storms, overwrite = TRUE)