+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `cur_data()` and `cur_data_all()` don't simplify list columns in rowwise data frames (#5901).
 
+* `storms` data updated to 2020 (@steveharoz, #5899).
+
 * `coalesce()` accepts 1-D arrays (#5557).
 
 * `filter()` forbids matrix results (#5973) and warns about data frame